requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cloudinary==1.28.0
Django==4.0
django-bootstrap4==21.1
django-crispy-forms==1.13.0
django-registration-redux==2.9
djangorestframework==3.12.4
psycopg2==2.9.2
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7
